refactor: Moved parser services into 'parsers' folder in core module.
