+++ src/app/core/core.module.ts
@@ -12,8 +12,8 @@ import { RundownEventParser } from './interfaces/rundown-event-parser.interface'
 import { BasicRundownStateService } from './services/basic-rundown-state.service'
 import { BasicRundownService } from './interfaces/basic-rundown-service.interface'
 import { EntityParser } from './interfaces/entity-parser.interface'
-import { ZodEntityParser } from './services/zod-entity-parser.service'
-import { ZodRundownEventParser } from './services/events/zod-rundown-event-parser.service'
+import { ZodEntityParser } from './parsers/zod-entity-parser.service'
+import { ZodRundownEventParser } from './parsers/zod-rundown-event-parser.service'
 import { EventSystemModule } from '../event-system/event-system.module'
 import { AdLibPieceService } from './interfaces/ad-lib-piece-service.interface'
